CON-56 [Back] 서비스간 통신 모델을 api 모듈에서 관리하도록 수정

+++ microservices/composite-service/src/main/java/com/connectcrew/teamone/compositeservice/composite/adapter/in/web/request/CreateRecruitRequest.java
@@ -1,7 +1,7 @@
 package com.connectcrew.teamone.compositeservice.composite.adapter.in.web.request;
 
+import com.connectcrew.teamone.api.projectservice.enums.MemberPart;
 import com.connectcrew.teamone.compositeservice.composite.application.port.in.command.CreateRecruitCommand;
-import com.connectcrew.teamone.compositeservice.composite.domain.enums.MemberPart;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -16,7 +16,7 @@ import lombok.Setter;
 public class CreateRecruitRequest {
     private MemberPart part;
     private String comment;
-    private Integer max;
+    private Long max;
 
     public CreateRecruitRequest(String json) {
         this.part = null;
@@ -25,7 +25,7 @@ public class CreateRecruitRequest {
     }
 
     @JsonCreator
-    public CreateRecruitRequest(@JsonProperty("part") String part, @JsonProperty("comment") String comment, @JsonProperty("max") int max) {
+    public CreateRecruitRequest(@JsonProperty("part") String part, @JsonProperty("comment") String comment, @JsonProperty("max") long max) {
         this.part = MemberPart.valueOf(part);
         this.comment = comment;
         this.max = max;
