+++ microservices/composite-service/src/main/java/com/connectcrew/teamone/compositeservice/composite/adapter/in/web/response/RecruitStatusResponse.java
@@ -8,8 +8,8 @@ public record RecruitStatusResponse(
         String part,
         String partKey,
         String comment,
-        Integer current,
-        Integer max,
+        Long current,
+        Long max,
         Boolean applied
 ) {
     public RecruitStatusResponse(RecruitStatus recruitStatus) {
